requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.2; python_version >= '3.7'
astroid==2.11.5; python_full_version >= '3.6.2'
autopep8==1.6.0
dill==0.3.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
django-cors-headers==3.13.0
django==4.0.5
djangorestframework==3.13.1
isort==5.10.1; python_version < '4.0' and python_full_version >= '3.6.1'
lazy-object-proxy==1.7.1; python_version >= '3.6'
mccabe==0.7.0; python_version >= '3.6'
pillow==9.1.1
platformdirs==2.5.2; python_version >= '3.7'
pycodestyle==2.8.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pylint-django==2.5.3
pylint-plugin-utils==0.7; python_full_version >= '3.6.2'
pylint==2.14.1
pytz==2022.1
setuptools==62.3.2; python_version >= '3.7'
sqlparse==0.4.2; python_version >= '3.5'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
tomli==2.0.1; python_version < '3.11'
tomlkit==0.11.0; python_version >= '3.6' and python_version < '4.0'
typing-extensions==4.2.0; python_version < '3.10'
wrapt==1.14.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
